// types/Types.ts
export interface SingleNoteType {
  _id: string;
  clerkUserId: string;
  title: string;
  isFavorite: boolean;
  tags: SingleTagType[];
  description: string;
  code: string;
  language: string;
  creationDate: string;
  isTrash: boolean;
}

export interface SingleTagType {
  _id: string;
  clerkUserId: string;
  name: string;
}